refactor the CO2 tables handling
this is ugly: the compile unit which features the main function now
needs to include the co2tables.inc.cc file in the appropriate
namespace. the reason for this is that the previous solution did not
work properly with slightly older compilers and the solution before
that did not work if the tables were used by multiple compile
units. since ewoms-material will be staying library-less, I went this
(ugly) route. suggestions for better approaches are highly
appreciated.